Refactor, split out presence broadcast plugin
